Refactored Example Structure
