+++ requirements.txt
@@ -1,3 +1,4 @@
+ipython
 pytest
 black
 coverage
