+++ tests/test_pythonpath.py
@@ -1,7 +1,7 @@
 """Test the PYTHONPATH sample."""
 import pytest
 
-from samples.beginner.pythonpath.app import display_pythonpath
+from samples.pythonpath.app import display_pythonpath
 
 
 def test_display_pythonpath():
